feat(response and params): response and parameters are not camlized now
BREAKING CHANGE: Now, parameters and responses are not formatted in casing. So, your parameters will
be sent as they are passed, and response returned will not be modified

BREAKING CHANGE: Now, parameters and responses are not formatted in casing. So, your parameters will
be sent as they are passed, and response returned will not be modified